capabilities: split client and server
